# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Pulse 109 Platform API
  version: 1.0.0
  description: >-
    Contract-first API for intake, operator assistance, incident management,
    analytics and regional system adapters. All timestamps use RFC 3339 with
    an explicit UTC offset. Human confirmation remains mandatory for routing,
    priority changes, duplicate merges and generated replies.
servers:
  - url: https://pulse109.example.gov.kz
    description: Production placeholder. Replace after infrastructure approval.
...
paths:
...
  /v1/requests:
    post:
      tags: [Requests]
      summary: Create or import an appeal
      operationId: createRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/RegionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequest'
      responses:
        '201':
          description: Appeal created
          headers:
            X-Correlation-Id:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appeal'
        '200':
          description: Existing appeal returned for a repeated idempotency key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appeal'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /v1/requests/{request_id}:
    get:
      tags: [Requests]
      summary: Get the current appeal snapshot and lifecycle timeline
      operationId: getRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Appeal snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppealDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/requests/{request_id}/classifications:
    post:
      tags: [Requests]
      summary: Produce a versioned AI recommendation
      operationId: classifyRequest
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClassificationInput'
      responses:
        '200':
          description: Recommendation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassificationRecommendation'
        '202':
          description: Request accepted for asynchronous inference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/requests/{request_id}/decisions:
    post:
      tags: [Decisions]
      summary: Confirm or correct the recommendation
      operationId: recordDecision
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperatorDecision'
      responses:
        '201':
          description: Human decision recorded in the audit trail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionReceipt'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
  /v1/requests/{request_id}/assignments:
    post:
      tags: [Decisions]
      summary: Assign or reassign an appeal to a service
      operationId: createAssignment
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignmentCommand'
      responses:
        '202':
          description: Assignment stored and queued for external synchronization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncReceipt'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1/requests/{request_id}/status-events:
    post:
      tags: [Requests]
      summary: Append a status event from an operator or source system
      operationId: appendStatusEvent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StatusEventInput'
      responses:
        '201':
          description: Event appended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LifecycleEvent'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /v1/reports:
    post:
      tags: [Reports]
      summary: Create a PDF or XLSX report asynchronously
      operationId: createReport
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportRequest'
      responses:
        '202':
          description: Report job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
  /v1/jobs/{job_id}:
    get:
      tags: [Operations]
      summary: Get an asynchronous job status
      operationId: getJob
      security:
        - bearerAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/RegionId'
      responses:
        '200':
          description: Current job state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/catalog/services:
    get:
      tags: [Catalog]
      summary: Get the versioned regional service catalog
      operationId: listServices
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RegionId'
        - name: effective_at
          in: query
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Effective service definitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServiceDefinition'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by the approved government identity provider.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    CorrelationId:
      name: X-Correlation-Id
      in: header
      required: false
      schema:
        type: string
        maxLength: 128
    RegionId:
      name: X-Region-Id
      in: header
      required: true
      schema:
        type: string
        pattern: '^[A-Z0-9_-]{2,32}$'
    RequestId:
      name: request_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    CorrelationId:
      schema:
        type: string
  responses:
    BadRequest:
      description: Malformed request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Version, idempotency or state conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Resource not found or not visible to this principal
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unprocessable:
      description: Business validation failed
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailable:
      description: Optional AI or retrieval dependency is unavailable
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    CreateRequest:
      type: object
      additionalProperties: false
      required:
        - source_system
        - source_request_id
        - region_id
        - received_at
        - received_at_quality
        - channel
      properties:
        source_system:
          type: string
          maxLength: 64
        source_request_id:
          type: string
          maxLength: 128
        region_id:
          type: string
        received_at:
          type: [string, 'null']
          format: date-time
        received_at_quality:
          type: string
          enum: [exact, source_tz_assumed, date_only, missing]
        source_timezone:
          type: [string, 'null']
        channel:
          type: string
          enum: [phone, web, mobile, telegram, whatsapp, email, walk_in, import, other]
        language:
          type: string
          enum: [kk, ru, mixed, unknown]
          default: unknown
        text:
          type: string
          maxLength: 20000
        transcript_ref:
          type: string
        media_refs:
          type: array
          maxItems: 20
          items:
            type: string
        location:
          $ref: '#/components/schemas/LocationInput'
        citizen_token:
          type: string
          description: Pseudonymous token. Direct identifiers are stored outside this API domain.
        source_payload_ref:
          type: string
        consent_or_legal_basis:
          type: string
          maxLength: 128
      allOf:
        - if:
            required: [received_at_quality]
            properties:
              received_at_quality:
                const: missing
          then:
            properties:
              received_at:
                type: 'null'
        - if:
            required: [received_at_quality]
            properties:
              received_at_quality:
                const: date_only
          then:
            properties:
              received_at:
                type: 'null'
        - if:
            required: [received_at_quality]
            properties:
              received_at_quality:
                const: exact
          then:
            required: [received_at]
            properties:
              received_at:
                type: string
                format: date-time
        - if:
            required: [received_at_quality]
            properties:
              received_at_quality:
                const: source_tz_assumed
          then:
            required: [received_at, source_timezone]
            properties:
              received_at:
                type: string
                format: date-time
              source_timezone:
                type: string
                minLength: 1
    LocationInput:
      type: object
      additionalProperties: false
      properties:
        address_text_private_ref:
          type: string
        geo_id:
          type: string
        object_id:
          type: string
        latitude:
          type: number
          minimum: -90
          maximum: 90
        longitude:
          type: number
          minimum: -180
          maximum: 180
        precision_m:
          type: number
          minimum: 0
    Appeal:
      allOf:
        - $ref: '#/components/schemas/CreateRequest'
        - type: object
          required: [request_id, created_at, version, status]
          properties:
            request_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            version:
              type: integer
              minimum: 1
            status:
              type: string
              enum: [new, triage, assigned, in_progress, resolved, closed, reopened, cancelled]
    AppealDetail:
      allOf:
        - $ref: '#/components/schemas/Appeal'
        - type: object
          properties:
            current_decision:
              $ref: '#/components/schemas/OperatorDecision'
            incident_id:
              type: [string, 'null']
              format: uuid
            timeline:
              type: array
              items:
                $ref: '#/components/schemas/LifecycleEvent'
            synchronization:
              $ref: '#/components/schemas/SyncState'
    ClassificationInput:
      type: object
      additionalProperties: false
      required: [request_version]
      properties:
        request_version:
          type: integer
          minimum: 1
        force_model_alias:
          type: string
          enum: [champion, challenger]
        return_similar:
          type: boolean
          default: false
    ClassificationRecommendation:
      type: object
      required:
        - recommendation_id
        - request_id
        - request_version
        - model_version
        - taxonomy_version
        - top_topics
        - top_services
        - priority
        - confidence_band
        - requires_human_confirmation
      properties:
        recommendation_id:
          type: string
          format: uuid
        request_id:
          type: string
          format: uuid
        request_version:
          type: integer
        model_version:
          type: string
        taxonomy_version:
          type: string
        top_topics:
          type: array
          minItems: 1
          maxItems: 3
          items:
            $ref: '#/components/schemas/RankedLabel'
        top_services:
          type: array
          minItems: 1
          maxItems: 3
          items:
            $ref: '#/components/schemas/RankedLabel'
        priority:
          type: string
          enum: [routine, elevated, urgent, emergency_handoff]
        confidence_band:
          type: string
          enum: [high, medium, low, out_of_domain]
        out_of_domain_score:
          type: number
          minimum: 0
          maximum: 1
        rule_hits:
          type: array
          items:
            type: string
        missing_fields:
          type: array
          items:
            type: string
        explanation:
          type: array
          items:
            type: string
        requires_human_confirmation:
          const: true
        produced_at:
          type: string
          format: date-time
    RankedLabel:
      type: object
      required: [id, score]
      properties:
        id:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 1
        display_name:
          type: string
    OperatorDecision:
      type: object
      additionalProperties: false
      required:
        - request_version
        - topic_id
        - service_id
        - priority
        - action
      properties:
        request_version:
          type: integer
          minimum: 1
        recommendation_id:
          type: [string, 'null']
          format: uuid
        topic_id:
          type: string
        service_id:
          type: string
        priority:
          type: string
          enum: [routine, elevated, urgent, emergency_handoff]
        action:
          type: string
          enum: [accepted, corrected, manual]
        correction_reason:
          type: [string, 'null']
          maxLength: 1000
        operator_note:
          type: [string, 'null']
          maxLength: 2000
    DecisionReceipt:
      type: object
      required: [decision_id, request_id, new_version, audit_event_id]
      properties:
        decision_id:
          type: string
          format: uuid
        request_id:
          type: string
          format: uuid
        new_version:
          type: integer
        audit_event_id:
          type: string
          format: uuid
    AssignmentCommand:
      type: object
      additionalProperties: false
      dependentRequired:
        expected_due_at: [policy_version]
      required: [request_version, service_id, reason_code]
      properties:
        request_version:
          type: integer
        service_id:
          type: string
        assignee_unit_id:
          type: string
        reason_code:
          type: string
        expected_due_at:
          type: [string, 'null']
          format: date-time
        policy_version:
          type: string
    StatusEventInput:
      type: object
      additionalProperties: false
      required: [source_event_id, status, occurred_at, occurred_at_quality, source_system]
      properties:
        source_event_id:
          type: string
          maxLength: 128
        status:
          type: string
          enum: [new, triage, assigned, accepted, in_progress, waiting, resolved, closed, reopened, cancelled]
        occurred_at:
          type: [string, 'null']
          format: date-time
        occurred_at_quality:
          type: string
          enum: [exact, source_tz_assumed, date_only, missing]
        source_timezone:
          type: [string, 'null']
        observed_at:
          type: string
          format: date-time
        source_system:
          type: string
        reason_code:
          type: string
        evidence_refs:
          type: array
          items:
            type: string
      allOf:
        - if:
            required: [occurred_at_quality]
            properties:
              occurred_at_quality:
                const: missing
          then:
            properties:
              occurred_at:
                type: 'null'
        - if:
            required: [occurred_at_quality]
            properties:
              occurred_at_quality:
                const: exact
          then:
            required: [occurred_at]
            properties:
              occurred_at:
                type: string
                format: date-time
        - if:
            required: [occurred_at_quality]
            properties:
              occurred_at_quality:
                const: source_tz_assumed
          then:
            required: [occurred_at, source_timezone]
            properties:
              occurred_at:
                type: string
                format: date-time
              source_timezone:
                type: string
                minLength: 1
    LifecycleEvent:
      type: object
      required: [event_id, event_type, occurred_at, observed_at, actor_type]
      properties:
        event_id:
          type: string
          format: uuid
        event_type:
          type: string
        occurred_at:
          type: [string, 'null']
          format: date-time
        occurred_at_quality:
          type: string
          enum: [exact, source_tz_assumed, date_only, missing]
        observed_at:
          type: string
          format: date-time
        actor_type:
          type: string
          enum: [citizen, operator, supervisor, service, source_system, model, rule, system]
        actor_id_token:
          type: string
        payload:
          type: object
          additionalProperties: true
...
    ReportRequest:
      type: object
      additionalProperties: false
      required: [format, template_id, query]
      properties:
        format:
          type: string
          enum: [pdf, xlsx]
        template_id:
          type: string
        query:
          $ref: '#/components/schemas/AnalyticsQuery'
        locale:
          type: string
          enum: [kk-KZ, ru-KZ]
    Job:
      type: object
      required: [job_id, type, status, created_at]
      properties:
        job_id:
          type: string
          format: uuid
        type:
          type: string
        status:
          type: string
          enum: [queued, running, succeeded, failed, cancelled]
        created_at:
          type: string
          format: date-time
        completed_at:
          type: [string, 'null']
          format: date-time
        result_ref:
          type: [string, 'null']
        error_code:
          type: [string, 'null']
    SyncReceipt:
      type: object
      required: [outbox_event_id, status]
      properties:
        outbox_event_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, delivered, retrying, failed_permanent]
        next_attempt_at:
          type: [string, 'null']
          format: date-time
...